move app submission tracker pillow to it's own module
